Turbopack Build: Fix dynamic-missing-gsp test (#79472)
## What?

Fixes the test by changing the path it checks, both webpack and
turbopack versions are okay, the Turbopack version is better though
holding the full path root to page.js file.

Also fixed the test to not add `export const dynamic = undefined` as
that is checked in Turbopack as it's not a valid string. This is also a
better error.