[build-sourcemaps] Increase stacktrace limit during prerender
Gated behind `experimental.enablePrerenderSourceMaps`.

A low limit increases the likelihood of the stacktrace
only including frames pointing into node_modules
which we consider unhelpful and ignore-list.

We already increase it to 50 in `next dev` so we should
do the same during prerender to avoid confusion.

I'll follow-up with checking the CPU/memory impact
of increasing this limit.
For debugging, it should really be +infinity.
Limiting the stacktrace to something useful is handled
by ignore-listing not an arbitrary number of frames.
Especially in the terminal, it's annoying to limit
it to N frames because you can't expand it later on
when you look at the error in the terminal.
Really, the terminal is just not a good debugging tool.