Turbopack Build: Fix next/font test (#79502)
## What?

Fixes the test which assumed specific hashes.

While digging into this with @sokra and @mischnic we found that the
reason it currently preloads 4 instead of 2 files is because of source
based side-effects detection that Turbopack does not have yet. We've
created an issue to track that so that this PR can land.