refactor(turbopack): Refactor usage of visitors (#79525)
### What?

We are using `VisitMut`, which is very heavy, to apply simple operations. This PR refactors it to use a custom trait, which is very lightweight and has a proper signature.

### Why?

For a small performance improvement and a smaller binary. I'll refactor to remove  `VisitorFactory` completely because we don't need such indirection anymore. This is possible because `AstModifier` takes `&self`, unlike `&mut self` of `VisitMut`.