[turbopack] Fix the stubbed out implementation of `signature`  so that it works in a web worker. (#79509)
## What
Fix the stubbed out runtime hook for react HMR to correctly support both the `register` and `signature` APIs, by moving the stubs into the runtime directly.  This way TSC can tell us we failed to uphold the function signature.

## Why

The `swc` react transform looks for certain naming conventions and function signatures to register top level items for HMR.  However, if a module is loaded in a web worker the react hooks are not installed.  We could potentially try to avoid running the react transform on modules loaded in workers but as a dev-only transform this might actually be slower (for shared modules).  So instead we just supply an alternate fake implementation.