[release-legacy] fix: set `stable` dist tag for backport releases instead of `latest` (#79596)
### Why?

If the current version is less than the latest, it means this is a
backport release. Since NPM sets the `latest` tag by default during
publishing, when users install `next@latest`, they might get the
backported version instead of the actual "latest" version. Therefore, we
explicitly set the tag as `stable` for backports.