Unflake `Undefined default export` test
[flakiness metric](https://app.datadoghq.com/ci/test/runs?query=test_level%3Atest%20env%3Aci%20%40git.repository.id%3Agithub.com%2Fvercel%2Fnext.js%20%40test.service%3Anextjs%20%40test.status%3Afail%20%40test.name%3A%22Undefined%20default%20export%20should%20error%20when%20page%20component%20export%20is%20not%20valid%22%20%40git.branch%3Acanary&agg_m=count&agg_m_source=base&agg_t=count&currentTab=overview&eventStack=&fromUser=false&index=citest&start=1746962081472&end=1748258081472&paused=false)

The test's flakiness started to increase significantly on 5/21.

Using
[these](https://github.com/vercel/next.js/actions/runs/15249566684/job/42888664439?pr=79565)
[two](https://github.com/vercel/next.js/actions/runs/15251316578/job/42888825737?pr=79629)
test runs as examples, we can see that the `/_error` page takes quite a
long time to compile, which then leads to a timeout while waiting for
the `#__next` element on the error page to be present in the DOM.

Looking at the canary release from this day, it might be related to #79260.

In fact, running the test before and after this PR uncovers a
significant increase in dev compilation times for the `/_error` page,
which is a pages router page. We'll address this regression in a
follow-up.