[ppr] Narrow condition for fallback shell generation at runtime (#79565)
This addresses a bug where a fallback shell was generated at runtime
when it was not expected to do so. This was an issue with how the
`x-now-route-matches` header was parsed and was incorrectly assuming
that a falsy header was the same as comparing it to a string. This
correctly only generates the fallback shell if the header is present and
was an empty string as was sent by Vercel.

Co-authored-by: Hendrik Liebau <mail@hendrik-liebau.de>