[dynamicIO] Only report client sync IO errors if they are above a Suspense boundary
Currently if the client has any sync IO during prerenders and no shell is produced dynamicIO will assume that the sync IO in the client is the cause. This is often incorrect because you might abort the prerender in a component inside a Suspense boundary but still have some reference to Request data that came from RSC that is not in a Suspense boundary. This misleads the user into thinking the sync IO is the problem when in fact it is a misplaced boundary around something else entirely.

This is a tricky one to solve and it may be that this doesn't fully cover all the cases it should. The technique is to abort on sync IO before stashing the sync IO error. All of the pending render tasks React has stashed away will abort and use normal processing ot decide if they violate the rules. Namely this is where messages like "A component accessed data, headers, params, searchParams, or a short-lived cache without a Suspense boundary" come from. Then the sync IO error is stashed. If any more (one more to be exact) render tasks abort after this we know that this task was the one that was currently rendering when the sync IO happened. If this is the case we include the sync IO error in the list of dynamic errors rather than the generic reason and only if it has no Suspense ancestor.

This fix is probably incomplete. For instance if you perform syncIO inside `queueMicrotask` then all of the render tasks will be aborted and we'll have no way to attribute it to any particular component stack and thus cannot know if it was inside or outside of a Suspense boundary. This will be solved (if possible) in a follow up change.