Implement handler interface for app-page (#79568)
Continuation of https://github.com/vercel/next.js/pull/78166 this
implements the handler interface for app router pages, This does not
move the response handling inside of the handler yet as that will be in
a follow-up PR to keep the changes isolated. This still returns the
RenderResult for the base-server to continue to handle.

Validated against deploy tests
https://github.com/vercel/vercel/pull/13389 and
https://github.com/vercel/next.js/actions/runs/15358616431/job/43222466204