Turbopack Build: Fix underscore path tests (#79778)
## What?

Implements handling for `%5F` to `_` conversion to handle pathnames in
App Router that want to bypass the underscore ignore folder.

Implementing this was a bit of a hassle not on the Turbopack side but
there's an additional layer of route handling before the compiler that
turns the filesystem paths into routable routes. Turns out there was
some additional places where the normalization needs to apply.

For now implemented it so that it special cases Turbopack handling,
might change the webpack implementation in the future to output the
files into the same path that Turbopack outputs it at. This is only
relevant for dev, production is handled the same already.