Display pid of process using port
E.g.:

```shell
  ⚠ Port 3000 is in use by process 3597, using available port 3004 instead.
```

I often find that I don't know process is using the port and sometimes
if I was running that process in an IDE and then the IDE crashed or
restarted, it may have left my process running somewhere and I prefer to
kill it and have a new one in my IDE, so I can watch log messages. This
makes it easier to do so, by telling the user the pid of the process
that is using the port. It still preserves the current behavior of
choosing a new port and starting on that port, but the user can opt to
abort the new server, kill the old process, and start the server again.