perf(turbopack): Use owned instance of `Code` for `minify()` (#79991)
### What?

Add `Rope::into_string()` that takes the owned instance of `self` and avoids an extra allocation.

### Why?

We don't need to go through the `Read` + `Clone` interface in these cases. We construct the `Code` instance, and pass it right away to the `minify` without cloning, so we can even use `String`.