[dev-tools] add restart dev server button to error overlay (#80060)
### Why?

When the user reloads on a specific error, and that error persists, we
show the restart server button as an option. This is because some errors
are recoverable by restarting the server and rebuilding the app.


https://github.com/user-attachments/assets/4e65d26c-8344-47b8-a6b9-1151008c271c

When Turbopack persistent cache is enabled, it will also clear the
bundler cache. This improves DX by replacing the need to run `rm -rf
.next` manually.


https://github.com/user-attachments/assets/e3f08176-5e9e-41fc-a8ce-9e731a45a6f2

Closes NEXT-4477

---------

Co-authored-by: Benjamin Woodruff <benjamin.woodruff@vercel.com>
Co-authored-by: Sebastian "Sebbie" Silbermann <sebastian.silbermann@vercel.com>