fix(turbo-persistence): Atomically mark old versioned directories for deletion before deleting them (#80082)
This uses a trick similar to what I did in https://app.graphite.dev/github/pr/vercel/next.js/79425 to mark the directory for deletion using an atomic operation (a directory rename in this case) before deleting it, so that we can resume the deletion upon restart if we're interrupted while deleting the directory.

This may be important as deleting large directories can take some time.

I think this is the source of these user reports: https://vercel.slack.com/archives/C046HAU4H7F/p1748616184765179