[dynamicIO] prioritize preprocessing RSC rows when prerendering (#80125)
When prerendering with dynamicIO we take advantage of a few things about
microtask behavior to ensure that anything prerenderable in RSC is
included in the HTML prerender. We found however that with large enough
RSC payloads certain things that should have been prerenderable could be
incorrectly excluded when paired with sync IO in the client because this
causes the prerender to abort earlier than one task. To prevent this we
allow the RSC response to pre-process entirely before continuing with
the prerender. This is possible as long as the RSC prerender is fully
complete before the HTML prerender begins.

We only do this when dynamicIO is enabled and only during prerender. By
construction this means that this also never happens in edge runtime.