Turbopack Build: Implement helpful error for missing sass package (#80155)
## What?

Implements the same error message that we add to the webpack handling. I'm not super happy about this solution (though it matches what we do for webpack) ideally we improve the sass-loader error itself in the future but that requires upgrading sass-loader as well and introducing a breaking change. For now this unblocks the test passing.