[dev-overlay] Move building indicator into Dev Overlay state (#79985)
Used to be a singleton treated as a sync external store. The singleton would require extra attention once we move into a separate bundle. Instead, I moved it into actual React state to get rid of the sync external store.
We want to avoid sync external stores as much as possible to leverage React's concurrent features to the fullest.

If we need to slice the state up, we can investigate that separately.