remove unique metadata prop from initial RSC payload (#79388)
This value is used as a key to re-mount `head` on subsequent client
navigations, but as far as I can tell, it's not needed for the initial
RSC payload that seeds these pages.

In its current form, these random IDs get serialized as part of the RSC
payload, leading to non-deterministic HTML outputs.

Fixes #77810