docs: fix grammar in Code of Conduct section ('them' → 'it')  (#80181)
<!-- Thanks for opening a PR! Your contribution is much appreciated.
To make sure your PR is handled as smoothly as possible we request that
you follow the checklist sections below.
Choose the right checklist for the change(s) that you're making:
-->

## What?

Fixed a grammatical issue in the Code of Conduct section of
`packages/next/README.md`.
Changed “adhere to them” to “adhere to it”.

## Why?

"Code of Conduct" is a singular noun, so using "them" is grammatically
incorrect. The correct pronoun is "it".

## How?

Edited the relevant sentence in `packages/next/README.md`.

<!--
Closes NEXT-
Fixes #
-->

Co-authored-by: JJ Kasper <jj@jjsweb.site>