Add a test to show that `'use cache'` works in `generateMetadata` (#80172)
When calling a function with a `'use cache'` directive from
`generateMetadata`, and from the page/layout, the invocations are
deduped and the result is shared.

When `generateMetadata` itself also has a `'use cache'` directive, and
it calls a nested function with a `'use cache'` directive, we do need to
land #78703 for proper deduplication of the nested cache.