Prerender with streaming metadata during revalidation (#80245)
During the export phase of `next build` we're hard-coding
`serveStreamingMetadata` to `true`, so we need to do the same during
revalidation.

Otherwise we're checking the user agent to decide if we should serve
streaming metadata.

This fixes a hydration error that occurred when hydrating the resumed
PPR shell. The error was not present for the initial resume of the
prerendered shell, and only after the shell was revalidated once.