[turbopack] Replace uses of `Value<Pattern>` with just `Pattern`  (#80134)
Replace uses of `Value<Pattern>` in the resolve crate with `Pattern`

This required a manual implementation of `TaskInput` which ended up being trivial since there are no captured `Vcs`. However, the `derive` macro failed because it is a recursive type.

### Why?
The Value<> type is confusing and error prone, lets destroy it!

As discussed over slack, turbo_tasks::Value type always implements is_resolved as true and is_transient as false which is incorrect and can enable smuggling transient vcs into turbo tasks.

• https://vercel.slack.com/archives/C03EWR7LGEN/p1743456121241529
• https://vercel.slack.com/archives/C04NGV98TPS/p1743455453764879