[test] Stop opening browser by default in local Dev Overlay Storybook (#80291)
This doesn't properly work with if a different Chrome release channel is used. It opens Chrome stable even if I have Beta as my default.
It's also annoying if you restart it and have a tab already open in you non-default browser.
We don't open a browser in Next.js for those same reasons.

Storybook will display a URL to click in the terminal so all this adds is one click if you start it for the first time.